+++ CHANGES.md
@@ -17,8 +17,9 @@ wavesurfer.js changelog
   - add `followCursorY` option (#1605)
 - Spectrogram plugin: fix `ready` listener when loading multiple
   audio files (#1572)
-- Regions plugin: fix `ready` listener when loading multiple
-  audio files (#1602)
+- Regions plugin:
+  - fix `ready` listener when loading multiple audio files (#1602)
+  - add `snapToGridInterval` and `snapToGridOffset` options
 - Timeline plugin: fix `ready` listener when loading multiple
   audio files
 - `util.getId()` now accepts a `prefix` argument (#1619)