+++ CHANGES.md
@@ -19,6 +19,7 @@ wavesurfer.js changelog
   audio files (#1572)
 - Regions plugin:
   - fix `ready` listener when loading multiple audio files (#1602)
+  - add `snapToGridInterval` and `snapToGridOffset` options (#1632)
   - allow drawing regions over existing regions, if the underlying ones are not
     draggable or resizable (#1633)
 - Timeline plugin: fix `ready` listener when loading multiple